`context="module"` で video 要素を定義してみた
React でも似たようなことは可能だったが、import タイミングで副作用を起こしたくなかったし、DI 的になってユニットテストがやりやすくなるかと思って、やっていなかった
Svelte は DOM 前提の世界だしユニットテストのための DI はそこまで考えなくてもいいかも、という気持ちでやってみた
