+++ src/main.ts
@@ -4,15 +4,12 @@ import { initializeFirestore } from "./firestore/initializeFirestore"
 import "./global.css"
 import { observeHash } from "./observeHash"
 import { keyWithDarkMode, observeDarkMode } from "./useDarkMode"
-import { createVideoTimer, keyWithVideoTimer } from "./useVideoTimer"
 
 run()
 
 async function run() {
   // https://neos21.net/blog/2018/08/19-01.html
   document.body.addEventListener("touchstart", () => {}, { passive: true })
-
-  const videoTimer = createVideoTimer()
 
   const darkMode$ = observeDarkMode()
 
@@ -27,7 +24,6 @@ async function run() {
     },
     context: new Map([
       ...firestoreImplContext(firestore),
-      keyWithVideoTimer(videoTimer),
       keyWithDarkMode(darkMode$),
     ]),
   })
